requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
Django==3.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
psycopg2==2.8.6
pycparser==2.21
pyparsing==3.0.9
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
